# v2/openapi/level5-v2.openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Level5.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v2/me/email/verification": {
      "post": {
        "tags": [
          "Account"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestEmailVerificationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestEmailVerificationDto"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/RequestEmailVerificationDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EmailVerificationDispatchResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailVerificationDispatchResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailVerificationDispatchResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmailVerificationDispatchResponseDto": {
        "type": "object",
        "properties": {
          "alreadyVerified": {
            "type": "boolean"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RequestEmailVerificationDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}